refactor: extract recall helpers from index.ts into tools/recall-helpers.ts
Extract buildRecallCandidateContext, collectLexicalRescueCandidates,
DiscoverTagStat, tokenizeTagDiscoveryText, countTokenOverlap,
escapeRegex, hasExactTagContextMatch, and PROJECT_SCOPE_BOOST constant
into a dedicated module. Update index.ts to import from the new module
and remove inline definitions, plus clean up now-unused imports
(computeRecallMetadataBoost, getSessionCachedProjectionTokens,
setSessionCachedProjectionTokens, tokenize, prepareTfIdfCorpusFromTokenizedDocuments,
rankDocumentsByTfIdf, LEXICAL_RESCUE_CANDIDATE_LIMIT/THRESHOLD/RESULT_LIMIT).
